+++ arangod/StorageEngine/PhysicalCollection.h
@@ -146,7 +146,7 @@ class PhysicalCollection {
   /// @param restore utilize specified ID, assume index has to be created
   virtual std::shared_ptr<Index> createIndex(
       velocypack::Slice info, bool restore, bool& created,
-      std::shared_ptr<std::function<arangodb::Result(uint64_t)>> = nullptr) = 0;
+      std::shared_ptr<std::function<arangodb::Result(double)>> = nullptr) = 0;
 
   virtual Result dropIndex(IndexId iid);
